[TASK] Allow execution of checks and quality tools on all branches
This change removes the branch contstrains for general
useful github workflowes, which helps contributers because
these tests can be run on forks. Thus allowing to have these
tests execute before opening a pull-request.

It is really annoying to open a pull request with a lot of
errors because it could not detected beforehand.